+++ R/get_table_odata4.R
@@ -1,6 +1,6 @@
-#' get_table_cbs_odata4 : retrieves CBS Odata4 table data
+#' Retrieving CBS Odata4 table data
 #'
-#' This function can be used to retrieve data or information from the new \href{https://www.cbs.nl}{CBS} ('Centraal Bureau voor de Statistiek' or 'Statistics Netherlands') Odata4 data infrastructure. The general OData4 protocol is described in \url{http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html} and the part that is implemented by CBS in  \url{https://acc-ccb.cbs.nl/implement.html}
+#' The function `get_table_cbs_odata4` can be used to retrieve data or information from the new \href{https://www.cbs.nl}{CBS} ('Centraal Bureau voor de Statistiek' or 'Statistics Netherlands') Odata4 data infrastructure. The general OData4 protocol is described in \url{http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html} and the part that is implemented by CBS in \url{https://beta-dataportal.cbs.nl/}
 #'
 #' @name get_table_cbs_odata4
 #' @param root Character string with the root for the url. When specified `odata_root` and `odata_cat` are not used
@@ -10,10 +10,9 @@
 #' @param verbose Boolean indicating if the generated url should be shown as a message.
 #' @param encode Boolean indicating if the query has to be encode by URLencode.
 #' @param odata_root Character string with the root for the url.
-#' @param odata_cat Character string with the catalog identifier.
-#' @param response Boolean indicating if the query has to return the httr response object instead of a table. Useful for debugging.
+#' @param odata_cat Character string with the catalog identifier.#' @param restype Character string indicating if the query has to return a `data.frame` or `list` (when `''`), a httr response object (when `resp`) or a JSON string (when `json`).
 #' @param error_msg Boolean indicating if the query has to return a more extensive error message. Useful for debugging.
-#' @return if not succesful a character string with an error message. If succesful a data.frame when the contents of a subtable was requested and a list with the properties of the table when Properties was requested. If `response==TRUE` , the result is a httr response object.
+#' @return if not succesful a character string with an error message. If succesful the contents is regarded as a json object and translated to a data.frame or list when possible. However if  `restype='resp'`  the result is a httr response object and when `restype='json'` the JSON string is not converted
 
 #' @export
 
@@ -28,6 +27,12 @@
 #' t1=get_table_cbs_odata4(table_id="82931NED",subtable="Observations",query="$skip=1&$top=2")
 #' t1=get_table_cbs_odata4(table_id="82931NED",subtable="Observations",query="$count")
 #' t1=get_table_cbs_odata4(table_id="82931NED",subtable="Observations",query="(7)")
+#'
+#' myurl = "https://beta-odata4.cbs.nl/CBS/82931NED/Observations?$format=json&$skip=1&$top=2"
+#' t1=get_table_cbs_odata4_GET(myurl)
+#' t1=get_table_cbs_odata4_GET(myurl,restype='resp')
+#' t1=get_table_cbs_odata4_GET(myurl,restype='json')
+#' t1=get_table_cbs_odata4_GET(myurl,error_msg=T)
 #' }
 #' @section Additional example:
 #' See \url{https://gist.github.com/HanOostdijk/787e9724dcd63735a431bcd16cbd18a0} for an analysis done with `get_table_cbs_odata4`. The resulting PDF file can be found in \url{https://www.hanoostdijk.nl/hugo/opendata_beta_versie4_dec2018_20181225.pdf} .
@@ -41,7 +46,7 @@ get_table_cbs_odata4 <-
 		encode = TRUE,
 		odata_root = "https://beta-odata4.cbs.nl",
 		odata_cat  = "CBS",
-		response = FALSE,
+		restype = '',
 		error_msg = TRUE) {
 		if (is.null(subtable)) {
 			subtable = ""
@@ -93,8 +98,8 @@ get_table_cbs_odata4 <-
 			display_wrapped(glue::glue("generated url  : {url1}"), w)
 			display_wrapped(glue::glue("unencoded query: {query}"), w)
 		}
-		res = get_table_cbs_odata4_GET(url1, response, error_msg)
-		if (post_code > 0 && !(class(res) == 'response')) {
+		res = get_table_cbs_odata4_GET(url1, restype, error_msg)
+		if (post_code > 0 && is.list(res) && class(res)!="response") {
 			res = as.data.frame((res)[-1])
 		}
 		res
@@ -102,27 +107,18 @@ get_table_cbs_odata4 <-
 
 #' get_table_cbs_odata4_GET : retrieves url
 #'
-#' This function handles the network IO for the \code{\link{get_table_cbs_odata4}} function but can also be used stand-alone.
+#' The function `get_table_cbs_odata4_GET` handles the network IO for the `get_table_cbs_odata4` function but can also be used stand-alone.
 #'
 #' @name get_table_cbs_odata4_GET
-#' @param url Character string with url that will be passed to GET function without further encoding. It is assumed that a json result can be returned. Default: none
-#' @param response Boolean indicating if the query has to return the httr response object instead of a table. Useful for debugging. Default: FALSE
-#' @param error_msg Boolean indicating if the query has to return a more extensive error message. Useful for debugging. Default: TRUE
-#' @return if not succesful a character string with an error message. If succesful the contents is regarded as a json object and translated to a data.frame or list when possible. If the parameter response=TRUE is set, the result is a httr response object.
+#' @param url Character string with url that will be passed to GET function without further encoding. It is assumed that a json result can be returned.
+#' @param restype Character string indicating if the query has to return a `data.frame` or `list` (when `''`), a httr response object (when `resp`) or a JSON string (when `json`).
+#' @param error_msg Boolean indicating if the query has to return a more extensive error message. Useful for debugging.
 
 #' @rdname get_table_cbs_odata4
 #' @export
 
-#' @examples
-#' \dontrun{
-#' myurl = "https://beta-odata4.cbs.nl/CBS/82931NED/Observations?$format=json&$skip=1&$top=2"
-#' t1=get_table_cbs_odata4_GET(myurl)
-#' t1=get_table_cbs_odata4_GET(myurl,response=T)
-#' t1=get_table_cbs_odata4_GET(myurl,error_msg=T)
-#' }
-
 get_table_cbs_odata4_GET <- function (url,
-	response = FALSE,
+	restype = '',
 	error_msg = TRUE) {
 	res1 = httr::GET(url)
 	err1 = httr::http_error(res1)
@@ -131,6 +127,7 @@ get_table_cbs_odata4_GET <- function (url,
 	cnt1 = httr::headers(res1)$`content-type`
 	jsn1 = stringr::str_detect(httr::headers(res1)$`content-type`, 'json')
 	xml1 = stringr::str_detect(httr::headers(res1)$`content-type`, 'xml')
+	restype = tolower(restype)
 	if (length(jsn1) == 0 && err1)
 		return(msg1)
 	if (jsn1) {
@@ -154,11 +151,13 @@ get_table_cbs_odata4_GET <- function (url,
 			}
 		}
 	}
-	if (response == TRUE) {
+	if (restype == 'resp') {
 		return(res1)
 	}
 	else if (err1)
 		return(msg1)
+	else if (restype == 'json')
+	  return(txt1)
 	else if (jsn1) {
 		if (!is.null(e$value)) {
 			return(e$value)